requirements.txt
pandas
numpy
requests
python-dotenv
schedule
matplotlib
python-binance
pytest